Project: 2 -- BlogSite/requirements.txt
Django==2.2.13
django-crispy-forms==1.9.1
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
